LayoutTests/fast/dom/HTMLElement/attr-true-string.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="div" contentEditable="true"></div>
<script>
description('When contentEditable attribute is "true" string, element.contentEditable returns "true" and the element is editable.')

shouldBe('document.getElementById("div").getAttribute("contentEditable")','"true"');
shouldBe('document.getElementById("div").contentEditable', '"true"');
shouldBe('document.getElementById("div").isContentEditable', 'true');
shouldBe('window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify")', '"read-write"');

document.getElementById("div").style.display= 'none';
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
